backend/requirements.txt
# requirements.txt
flask
flask_sqlalchemy
pymysql
flask_cors
faker
python-dotenv